+++ shuttlelog/src/main/java/edu/bsu/shuttlelog/dao/LogDAO.java
@@ -1,6 +1,5 @@
 package edu.bsu.shuttlelog.dao;
 
-import java.math.BigInteger;
 import java.util.List;
 
 import edu.bsu.shuttlelog.entity.Log;
@@ -13,12 +12,12 @@ public interface LogDAO {
 	public Log getByID(long id);
 
 
-	public List<BigInteger> save(List<Log> log);
+	public List<Long> save(List<Log> log);
 
 
 	public void update(long id, Log log);
 
 
-	BigInteger save(Log log);
+	long save(Log log);
 
 }
