+++ shuttlelog/src/main/java/edu/bsu/shuttlelog/dao/LoopDAO.java
@@ -1,6 +1,5 @@
 package edu.bsu.shuttlelog.dao;
 
-import java.math.BigInteger;
 import java.util.List;
 
 import edu.bsu.shuttlelog.entity.Loop;
@@ -11,8 +10,7 @@ public interface LoopDAO {
 
 	public Loop getByName(String loopName) throws Exception;
 
-	public BigInteger save(Loop loop) throws Exception;
+	public long save(Loop loop) throws Exception;
 
 	public Loop update(String loopName, Loop loop) throws Exception;
-
 }
